# Added a specialized implementation of WriteRLEBitPackedHybrid for int32s being written as int64