This case is used in writing the definition levels and repetition levels for each page, and thus is performance sensitive.

```
name         old time/op    new time/op    delta
WriteCSV-44    6.19ms ± 2%    5.32ms ± 2%  -14.01%  (p=0.000 n=20+17)

name         old alloc/op   new alloc/op   delta
WriteCSV-44    5.44MB ± 0%    4.44MB ± 0%  -18.44%  (p=0.000 n=19+18)

name         old allocs/op  new allocs/op  delta
WriteCSV-44     61.5k ± 0%     21.4k ± 0%  -65.24%  (p=0.000 n=20+20)
```