@alex 2020-05-13
Updated measurement:

```
name         old time/op    new time/op    delta
WriteCSV-44    5.96ms ±11%    4.71ms ±11%  -21.07%  (p=0.000 n=18+20)

name         old alloc/op   new alloc/op   delta
WriteCSV-44    5.44MB ± 0%    4.44MB ± 0%  -18.44%  (p=0.000 n=18+18)

name         old allocs/op  new allocs/op  delta
WriteCSV-44     61.5k ± 0%     21.4k ± 0%  -65.24%  (p=0.000 n=20+20)
```