@xitongsys 2020-05-15
hi, @alex 
Actually I don't remember why I did convert int32 to int64 for writing. 
I should investigate it...