@alex 2020-05-15
FWIW, even without that, I think you'd still want a specialized implementation so that we don't have to build a `[]interface{}` from an `[]int32`