@alex 2020-05-16
After thinking about it some more, I don't think you ever need the int64 for writing. I'll update the PR accordingly.